feat(sdk-core): switch tss signing b/w mpc v1/v2
TICKET: WP-1680
