+++ modules/sdk-core/src/bitgo/utils/tss/ecdsa/ecdsa.ts
@@ -3,17 +3,7 @@ import { Buffer } from 'buffer';
 import * as openpgp from 'openpgp';
 import { Key, SerializedKeyPair } from 'openpgp';
 import { Hash } from 'crypto';
-import createKeccakHash from 'keccak';
-
-import {
-  DklsDsg,
-  EcdsaPaillierProof,
-  EcdsaRangeProof,
-  EcdsaTypes,
-  hexToBigInt,
-  minModulusBitLength,
-  DklsTypes,
-} from '@bitgo/sdk-lib-mpc';
+import { EcdsaPaillierProof, EcdsaRangeProof, EcdsaTypes, hexToBigInt, minModulusBitLength } from '@bitgo/sdk-lib-mpc';
 import { bip32 } from '@bitgo/utxo-lib';
 
 import { ECDSA, Ecdsa } from '../../../../account-lib/mpc/tss';
@@ -40,8 +30,8 @@ import {
   TSSParamsForMessage,
   TxRequest,
 } from '../baseTypes';
-import { getTxRequest, sendSignatureShare } from '../../../tss';
-import { AShare, DShare, EncryptedNShare, OShare, SendShareType, SShare, WShare } from '../../../tss/ecdsa/types';
+import { getTxRequest } from '../../../tss';
+import { AShare, DShare, EncryptedNShare, SendShareType, SShare, WShare, OShare } from '../../../tss/ecdsa/types';
 import { createShareProof, generateGPGKeyPair, getBitgoGpgPubKey } from '../../opengpgUtils';
 import { BitGoBase } from '../../../bitgoBase';
 import { BackupProvider } from '../../../wallet';
@@ -55,15 +45,6 @@ import {
   TxRequestChallengeResponse,
 } from '../../../tss/types';
 import { BaseEcdsaUtils } from './base';
-import {
-  getSignatureShareRoundOne,
-  getSignatureShareRoundThree,
-  getSignatureShareRoundTwo,
-  verifyBitGoMessagesAndSignaturesRoundOne,
-  verifyBitGoMessagesAndSignaturesRoundTwo,
-} from '../../../tss/ecdsa/dkls';
-
-import { MPCv2SignatureShareRound1Output, MPCv2SignatureShareRound2Output } from '@bitgo/public-types';
 
 const encryptNShare = ECDSAMethods.encryptNShare;
 
@@ -971,171 +952,6 @@ export class EcdsaUtils extends BaseEcdsaUtils {
     return await getTxRequest(this.bitgo, this.wallet.id(), txRequest.txRequestId);
   }
 
-  private async signTxRequestMpcV2(
-    params: TSSParams | TSSParamsForMessage,
-    requestType: RequestType
-  ): Promise<TxRequest> {
-    console.time('SDK-WP-HSM Full back and forth signing');
-    const userKeyShare = Buffer.from(params.prv, 'base64');
-    console.time('Getting first tx request with txn info');
-    const txRequest: TxRequest =
-      typeof params.txRequest === 'string'
-        ? await getTxRequest(this.bitgo, this.wallet.id(), params.txRequest)
-        : params.txRequest;
-
-    console.timeEnd('Getting first tx request with txn info');
-    let derivationPath = '';
-    let msgToSign = '';
-    const userGpgKey = await generateGPGKeyPair('secp256k1');
-    const bitgoGpgPubKey = (await getBitgoGpgPubKey(this.bitgo)).mpcV2;
-    if (!bitgoGpgPubKey) {
-      throw new Error('Missing BitGo GPG key for MPCv2');
-    }
-
-    if (requestType === RequestType.tx) {
-      assert(txRequest.transactions || txRequest.unsignedTxs, 'Unable to find transactions in txRequest');
-      const unsignedTx =
-        txRequest.apiVersion === 'full' ? txRequest.transactions![0].unsignedTx : txRequest.unsignedTxs[0];
-      msgToSign = unsignedTx.signableHex;
-      derivationPath = unsignedTx.derivationPath;
-    } else if (requestType === RequestType.message) {
-      throw new Error('DKLS message signing not supported yet.');
-    }
-
-    let hash: Hash;
-    try {
-      hash = this.baseCoin.getHashFunction();
-    } catch (err) {
-      hash = createKeccakHash('keccak256') as Hash;
-    }
-    console.time(
-      'First round time Creating first round message from user, Sending Request And Waiting for HSM Response'
-    );
-    const hashBuffer = hash.update(Buffer.from(msgToSign, 'hex')).digest();
-    // const hashBufferHex = Buffer.from(hashBuffer).toString('hex');
-    const otherSigner = new DklsDsg.Dsg(userKeyShare, 0, derivationPath, hashBuffer);
-    const userSignerBroadcastMsg1 = await otherSigner.init();
-    const signatureShareRound1 = await getSignatureShareRoundOne(userSignerBroadcastMsg1, userGpgKey);
-    await sendSignatureShare(
-      this.bitgo,
-      txRequest.walletId,
-      txRequest.txRequestId,
-      signatureShareRound1,
-      RequestType.tx,
-      undefined,
-      'ecdsa',
-      'full',
-      userGpgKey.publicKey
-    );
-    console.timeEnd(
-      'First round time Creating first round message from user, Sending Request And Waiting for HSM Response'
-    );
-    console.time('First round time Getting TxRequest with HSMs result');
-    let latestTxRequest = await getTxRequest(this.bitgo, txRequest.walletId, txRequest.txRequestId);
-    console.timeEnd('First round time Getting TxRequest with HSMs result');
-    assert(latestTxRequest.transactions);
-    console.time(
-      'Second round time Creating second round message from user, Sending Request And Waiting for HSM Response'
-    );
-    const bitgoToUserMessages1And2 = latestTxRequest.transactions[0].signatureShares;
-    // TODO: Use codec for parsing
-    const parsedBitGoToUserSigShareRoundOne = JSON.parse(
-      bitgoToUserMessages1And2[bitgoToUserMessages1And2.length - 1].share
-    ) as MPCv2SignatureShareRound1Output;
-    if (parsedBitGoToUserSigShareRoundOne.type !== 'round1Output') {
-      throw new Error('Unexpected signature share response. Unable to parse data.');
-    }
-    const serializedBitGoToUserMessagesRound1And2 = await verifyBitGoMessagesAndSignaturesRoundOne(
-      parsedBitGoToUserSigShareRoundOne,
-      userGpgKey,
-      bitgoGpgPubKey
-    );
-
-    /** Round 2 **/
-    const deserializedMessages = DklsTypes.deserializeMessages(serializedBitGoToUserMessagesRound1And2);
-    const userToBitGoMessagesRound2 = otherSigner.handleIncomingMessages({
-      p2pMessages: [],
-      broadcastMessages: deserializedMessages.broadcastMessages,
-    });
-    const userToBitGoMessagesRound3 = otherSigner.handleIncomingMessages({
-      p2pMessages: deserializedMessages.p2pMessages,
-      broadcastMessages: [],
-    });
-    const signatureShareRoundTwo = await getSignatureShareRoundTwo(
-      userToBitGoMessagesRound2,
-      userToBitGoMessagesRound3,
-      userGpgKey,
-      bitgoGpgPubKey
-    );
-    await sendSignatureShare(
-      this.bitgo,
-      txRequest.walletId,
-      txRequest.txRequestId,
-      signatureShareRoundTwo,
-      RequestType.tx,
-      undefined,
-      'ecdsa',
-      'full',
-      userGpgKey.publicKey
-    );
-    console.timeEnd(
-      'Second round time Creating second round message from user, Sending Request And Waiting for HSM Response'
-    );
-    console.time('Second round time Getting HSMs response');
-    latestTxRequest = await getTxRequest(this.bitgo, txRequest.walletId, txRequest.txRequestId);
-    console.timeEnd('Second round time Getting HSMs response');
-    assert(latestTxRequest.transactions);
-    const txRequestSignatureShares = latestTxRequest.transactions[0].signatureShares;
-    // TODO: Use codec for parsing
-    const parsedBitGoToUserSigShareRoundTwo = JSON.parse(
-      txRequestSignatureShares[txRequestSignatureShares.length - 1].share
-    ) as MPCv2SignatureShareRound2Output;
-    const serializedBitGoToUserMessagesRound3 = await verifyBitGoMessagesAndSignaturesRoundTwo(
-      parsedBitGoToUserSigShareRoundTwo,
-      userGpgKey,
-      bitgoGpgPubKey
-    );
-    console.time(
-      'Third round time Creating final message from user and Sending request to WP for combining signatures'
-    );
-    /** Round 3 **/
-    const deserializedBitGoToUserMessagesRound3 = DklsTypes.deserializeMessages({
-      p2pMessages: serializedBitGoToUserMessagesRound3.p2pMessages,
-      broadcastMessages: [],
-    });
-    const userToBitGoMessagesRound4 = otherSigner.handleIncomingMessages({
-      p2pMessages: deserializedBitGoToUserMessagesRound3.p2pMessages,
-      broadcastMessages: [],
-    });
-
-    const signatureShareRoundThree = await getSignatureShareRoundThree(
-      userToBitGoMessagesRound4,
-      userGpgKey,
-      bitgoGpgPubKey
-    );
-    // Submit for final signature share combine
-    await sendSignatureShare(
-      this.bitgo,
-      txRequest.walletId,
-      txRequest.txRequestId,
-      signatureShareRoundThree,
-      RequestType.tx,
-      undefined,
-      'ecdsa',
-      'full',
-      userGpgKey.publicKey
-    );
-    console.timeEnd(
-      'Third round time Creating final message from user and Sending request to WP for combining signatures'
-    );
-
-    console.time('Third round time getting fully signed tx request');
-    const finalRequest = await getTxRequest(this.bitgo, this.wallet.id(), txRequest.txRequestId);
-    console.timeEnd('Third round time getting fully signed tx request');
-    console.timeEnd('SDK-WP-HSM Full back and forth signing');
-    return finalRequest;
-  }
-
   /**
    * Signs the transaction associated to the transaction request.
    * @param {string | TxRequest} params.txRequest - transaction request object or id
@@ -1145,7 +961,7 @@ export class EcdsaUtils extends BaseEcdsaUtils {
    */
   async signTxRequest(params: TSSParams): Promise<TxRequest> {
     this.bitgo.setRequestTracer(params.reqId);
-    return this.signTxRequestMpcV2(params, RequestType.tx);
+    return this.signRequestBase(params, RequestType.tx);
   }
 
   /**
